Add Android 16KB page size support and update Flutter version requirements
  - Update minimum Flutter version to 3.24.0 and recommended to 3.27.0
  - Upgrade build tools for 16KB compatibility:
    * Android Gradle Plugin (AGP) to 8.9.0
    * Gradle to 8.10/8.11.1
    * Kotlin to 2.1.10
    * NDK to r28 (28.0.12674087)
  - Update minimum Android API level from 21 to 24
  - Update recommended Android API level from 34 to 35
  - Migrate to 64-bit only architectures (arm64-v8a, x86_64)
  - Update iOS platform to 16.0
  - Add comprehensive Android Build Requirements documentation
  - Update all sample apps with consistent build configurations
  - Update dependencies across packages and examples

  This ensures compliance with Google Play's 16KB page size requirement
  (effective November 1, 2025) and provides optimal performance on modern
  Android devices.

+++ packages/hms_room_kit/lib/src/widgets/chat_widgets/pin_chat_widget.dart
@@ -113,8 +113,8 @@ class _PinChatWidgetState extends State<PinChatWidget> {
                                         mainAxisSize: MainAxisSize.max,
                                         dotsCount: data.item2,
                                         position: currentPage >= data.item2
-                                            ? 0
-                                            : currentPage,
+                                            ? 0.0
+                                            : currentPage.toDouble(),
                                         decorator: DotsDecorator(
                                           spacing: const EdgeInsets.only(
                                               bottom: 3.0, right: 8),
