+++ packages/hms_room_kit/lib/src/widgets/meeting_modes/custom_one_to_one_grid.dart
@@ -124,7 +124,7 @@ class _CustomOneToOneGridState extends State<CustomOneToOneGrid> {
                                 child: DotsIndicator(
                                   dotsCount: pageCount,
                                   position:
-                                      currentPage > pageCount ? 0 : currentPage,
+                                      currentPage > pageCount ? 0.0 : currentPage.toDouble(),
                                   decorator: DotsDecorator(
                                       activeColor:
                                           HMSThemeColors.onSurfaceHighEmphasis,
